refactor(nft): extract useMasterGardeningSchoolClaimStatus hook
The 'can claim this NFT?' check was duplicated in four wrappers — the
Collectibles and Market NftCard/NftTable components each carried their
own copy of the same useEffect + setState + handleClaim pattern, with
slightly different deps (some included 'identifier', some 'setIsClaimable').

Pull the pattern into one hook at src/hooks/useMasterGardeningSchoolClaimStatus.ts
that takes a variationId, exposes { isClaimable, handleClaim, refresh },
and centralises the account-and-team gating. Callers collapse from ~30
lines of polling boilerplate to one hook call.

Note: the TODOs in ClaimNftModal.tsx and Mint.tsx target a separate
concern (PLANT allowance check) and are left for a follow-up.
